homeassistant-addon/requirements.txt
# Production dependencies for ha-mcp
# Generated from pyproject.toml for Docker deployment

fastmcp>=2.11.0
httpx>=0.27.0
fuzzywuzzy>=0.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
websockets>=12.0
python-levenshtein>=0.25.0
